OpenTelemetryInstrumentationBase

OpenTelemetryInstrumentationBase.init — Method

Schema

Meter Name	Instrument Name	Instrument Type	Unit	Description

Base.Sys	uptime	ObservableGauge	second	The current system uptime in seconds.

Base.Sys	free_memory	ObservableGauge	bytes	The total free memory in RAM in bytes.

Base.Sys	free_memory_ratio	ObservableGauge	%	The ratio of free memory in percentage.

Base.Sys	maxrss	ObservableGauge	bytes	The maximum resident set size utilized in bytes.The current system uptime in seconds.

Base	jit_total_bytes	ObservableGauge	bytes	The total amount (in bytes) allocated by the just-in-time compiler.

Base	gc_live_bytes	ObservableGauge	bytes	The total size of live objects after the last garbage collection, plus the number of bytes allocated since then.

Base	gc_time_ns	ObservableGauge	nanoseconds	Total time spend in garbage collection, in nanoseconds