OpenTelemetrySDK

Exporters

Two common exporters are provided to for debugging:

OpenTelemetrySDK.AbstractExporter — Type

An AbstractExporter is to export a collection of AbstractSpans and (or) Metrics. Each method should have the following interfaces implemented:

force_flush!(::AbstractExporter)

shut_down!(::AbstractExporter)

export!(::AbstractExporter, collection)

OpenTelemetrySDK.ConsoleExporter — Type

ConsoleExporter(;io=stdout)

Print an AbstractSpan or Metric into io.

OpenTelemetrySDK.InMemoryExporter — Type

InMemoryExporter(;pool=[], is_shut_down=Ref(false))

Simply store all export!ed elements into the pool.

Tracer

TracerProvider

In SDK, TracerProvider and Span are provided to replace the dummy ones in API. Span is not exported since we mainly use with_span to create new spans.

OpenTelemetrySDK.TracerProvider — Type

TracerProvider(;kw...)

Keyword Arguments

sampler::AbstractSampler=DEFAULT_ON

resource=Resource()

span_processor::AbstractSpanProcessor=CompositSpanProcessor()

id_generator::AbstractIdGenerator=RandomIdGenerator()

limit_info=LimitInfo()

The following extra methods are provided beyond those defined in AbstractTracerProvider:

force_flush!(p::TracerProvider)

shut_down!(p::TracerProvider)

Base.push!(p::TracerProvider, sp::AbstractSpanProcessor)

OpenTelemetrySDK.force_flush! — Method

force_flush!(p::TracerProvider)

Shorthand to force flush inner span processors

OpenTelemetrySDK.shut_down! — Method

shut_down!(p::TracerProvider)

Shut down inner span processors and then mark itself as shut down.

ID Generators

OpenTelemetrySDK.RandomIdGenerator — Type

RandomIdGenerator(rng=Random.GLOBAL_RNG)

Use the rng to generate a random trace id or span id.

Samplers

OpenTelemetrySDK.ALWAYS_OFF — Constant

Always drop the span.

OpenTelemetrySDK.ALWAYS_ON — Constant

Always sample the span.

OpenTelemetrySDK.DEFAULT_OFF — Constant

Sampler that respects its parent span's sampling decision, but otherwise never samples.

OpenTelemetrySDK.DEFAULT_ON — Constant

Sampler that respects its parent span's sampling decision, but otherwise always samples.

Span Processors

OpenTelemetrySDK.AbstractSpanProcessor — Type

Each span processor must implement the following methods:

on_start!(span_processor::AbstractSpanProcessor, span::AbstractSpan)

on_end!(span_processor::AbstractSpanProcessor, span::AbstractSpan)

shut_down!(span_processor::AbstractSpanProcessor)

force_flush!(span_processor::AbstractSpanProcessor)

OpenTelemetrySDK.CompositSpanProcessor — Type

CompositSpanProcessor(processors...)

A wrapper of different concrete span processors. Users can also push! new span processors into it after construction.

See also SimpleSpanProcessor.

OpenTelemetrySDK.SimpleSpanProcessor — Type

SimpleSpanProcessor(span_exporter)

Export each span immediately when on_end! is called on this processor.

Metric

The current implementation of metrics in SDK is mainly inspired by the dotnet sdk.

┌──────────────────────────────────────────┐
│MeterProvider                             │
│                                          │
│  meters                                  │
│  views                                   │
│                                          │
│  instrument_associated_metric_names      │
│    instrument =>  Set{metric_name}       │
│                                          │
│  metrics                                 │
│    name => metric                        │
│    ┌───────────────────────────────────┐ │
│    │Metric                             │ │
│    │                                   │ │
│    │  name                             │ │
│    │  description                      │ │
│    │  criteria                         │ │
│    │  aggregation                      │ │
│    │    ┌──────────────────────────┐   │ │
│    │    │AggregationStore          │   │ │
│    │    │                          │   │ │
│    │    │  attributes => data_point│   │ │
│    │    │   ┌─────────────────┐    │   │ │
│    │    │   │AbstractDataPoint│    │   │ │
│    │    │   │                 │    │   │ │
│    │    │   │  value          │    │   │ │
│    │    │   │  start_time     │    │   │ │
│    │    │   │  end_time       │    │   │ │
│    │    │   │  exemplars      │    │   │ │
│    │    │   │ ┌────────────┐  │    │   │ │
│    │    │   │ │Exemplar    │  │    │   │ │
│    │    │   │ │            │  │    │   │ │
│    │    │   │ │ value      │  │    │   │ │
│    │    │   │ │ trace_id   │  │    │   │ │
│    │    │   │ │ span_id    │  │    │   │ │
│    │    │   │ └────────────┘  │    │   │ │
│    │    │   │                 │    │   │ │
│    │    │   └─────────────────┘    │   │ │
│    │    │                          │   │ │
│    │    └──────────────────────────┘   │ │
│    │                                   │ │
│    └───────────────────────────────────┘ │
│                                          │
└──────────────────────────────────────────┘

A View specifies which instruments are grouped together through Criteria. For each view, a Metric is created to store the Measurements. Each metric may have many different dimensions configured by StaticAttrs in a Measurement. For each dimension, we may also collect those Exemplars in the mean while.

Design decisions

For each registered instrument, we have stored the associated metrics configured by views into the instrument_associated_metric_names field. So that for each pair of instrument => measurement, we can quickly determine which metrics to update.

To make sure that measurements of the same attribute key-values but with different order can be updated in the same dimension in the AggregationStore, a design from opentelemetry-dotnet#2374 is borrowed here.

OpenTelemetrySDK.DROP — Constant

All measurement will be dropped.

OpenTelemetrySDK.Exemplar — Type

Exemplar(;kw...)

Exemplars are example data points for aggregated data. Read the specification to understand its relation to trace and metric.

Keyword arguments:

value

time_unix_nano

filtered_attributes::StaticAttrs, extra attributes of a Measurement that are not included in a Metric's :attribute_keys field.

trace_id, the trace_id in the span context when the measurement happens.

span_id, the span_id in the span context when the measurement happens.

OpenTelemetrySDK.HistogramAgg — Type

HistogramAgg(args...)

Arguments:

boundaries::NTuple{M, Float64} where M, the boundaries to calculate histogram buckets. Note that -Inf and Inf shouldn't be included.

is_record_min

is_record_max

agg_store

exemplar_reservoir_factory, when set to nothing, no exemplar will be stored.

OpenTelemetrySDK.HistogramAgg — Method

HistogramAgg{T}(;boundaries = DEFAULT_HISTOGRAM_BOUNDARIES, is_record_min = true, is_record_max = true)

OpenTelemetrySDK.LastValueAgg — Type

LastValueAgg(agg_store::AggregationStore, exemplar_reservoir_factory)

When exemplar_reservoir_factory set to nothing, no exemplar will be stored.

See more details in the specification.

OpenTelemetrySDK.LastValueAgg — Method

LastValueAgg{T}()

OpenTelemetrySDK.SumAgg — Type

SumAgg(agg_store::AggregationStore, exemplar_reservoir_factory)

When exemplar_reservoir_factory set to nothing, no exemplar will be stored. See more details in the specification.

OpenTelemetrySDK.SumAgg — Method

SumAgg{T}()

OpenTelemetrySDK.MeterProvider — Method

MeterProvider(;resource = Resource(), views = View[], n_max_metrics = N_MAX_METRICS)

If views is empty, a default one (View(;instrument_name="*") will be added to enable all metrics.

OpenTelemetrySDK.View — Type

View(name=nothing;kwargs...)

The name support wildcard.

See more details in the specification.

Keyword arguments

description = nothing,

attribute_keys = nothing,

extra_dimensions = StaticAttrs(),

aggregation = nothing,

instrument_name = nothing,

instrument_type = nothing,

meter_name = nothing,

meter_version = nothing,

meter_schema_url = nothing,

Misc

OpenTelemetrySDK.AbstractIdGenerator — Type

Customized id generators must implement the following two methods:

generate_span_id(::AbstractIdGenerator)

generate_trace_id(::AbstractIdGenerator)

OpenTelemetrySDK.AbstractMetricReader — Type

All metric readers should implement shut_down!(::AbstractMetricReader) and (r::AbstractMetricReader)().

Builtin readers:

CompositMetricReader

MetricReader

PeriodicMetricReader

OpenTelemetrySDK.AbstractSampler — Type

A sampler controls whether to drop a span or not when creating new spans in a certain context. Each sampler should have the should_sample method implemented.

OpenTelemetrySDK.AggregationStore — Method

AggregationStore{D}(;kw...) where D<:DataPoint

The AggregationStore holds all the aggregated datapoints in a Metric.

Keyword arguments

n_max_points = N_MAX_POINTS_PER_METRIC, the maximum number of data points.

n_max_attrs = 2 * N_MAX_POINTS_PER_METRIC, the maximum number of allowed attributes. Note that each datapoint may have several attributes pointing to them. (Those attributes have the same key-value pair but with different order)

OpenTelemetrySDK.MetricReader — Type

MetricReader([global_meter_provider()], [ConsoleExporter()])

Note that all metrics will be read on initialization.

OpenTelemetrySDK.MetricReader — Method

(r::MetricReader)()

For async instruments in r, their callbacks will be executed first before reading all the metrics.

OpenTelemetrySDK.PeriodicMetricReader — Type

PeriodicMetricReader(reader; export_interval_seconds = 60, export_timeout_seconds = 30)

Periodically call the reader in the background.

OpenTelemetrySDK.PeriodicMetricReader — Method

For PeriodicMetricReader, there's no need to call this method since the read operation will be done periodically in the background.

Base.push! — Method

Base.push!([p::TracerProvider], sp::AbstractSpanProcessor)

Add an extra span processor sp into the TracerProvider p.

OpenTelemetrySDK.on_end! — Method

on_end!(ssp::SimpleSpanProcessor, span)

The span is exported immediately if it is sampled.

OpenTelemetrySDK.on_start! — Method

on_start!(ssp::SimpleSpanProcessor, span)

OpenTelemetrySDK.should_sample — Function

should_sample(s::AbstractSampler, args...)

Arguments

parent_context::Context,

trace_id::TraceIdType,

name::String, the span name

kind::SpanKind,

attributes, StaticAttrs or DynamicAttrs

links, vector of Link

trace_state::TraceState,