Design

The content in this page is organized in the same order as the OpenTelemetry Specification.

API

Context

Context is implemented as a wrapper of NamedTuple, which means it is immutable. Each Task has exactly ONE Context instance, which is injected into the task_local_storage of the current_task by the parent task automatically.

Warning

Type piracy is used to the propagate context between tasks.

create_key is used to create a context key. But it is not exported yet because it seems to be only used internally until now.

Base.getindex(::Context, key) is implemented so to get a value in a Context, one can simply call ctx[key] to get the associated value of a key in a ctx.

Setting value of a Context is not directly supported. Given that Context is immutable, updating an immutable object in Julia seems strange. We provide the with_context function to create a new context based on the key-value pairs in the current_context. This syntax is more common than the attach/detach operations in the original specification.

OpenTelemetryAPI.with_context — Function

with_context(f, [context]; kv...)

Run function f in the context. If extra kv pairs are provided, they will be merged with the context to form a new context. When context is not provided, the current_context will be used.

OpenTelemetryAPI.current_context — Function

Return the Context associated with the caller's current execution unit.

Propagators

inject! and extract are provided to

The global propagator can be set to a CompositePropagator, with multiple dispatch, each inner propagator can be customized to handle different contexts and carriers.

TextMapPropagator is not implemented yet! Personally I feel that every propagator may depends on a third party package. To minimize the dependencies of OpenTelemetryAPI.jl, those specialized propagators can be registered as independent packages.

OpenTelemetryAPI.inject! — Function

inject(carrier, [global_propagator], [current_context])

Injects the value into a carrier. For example, into the headers of an HTTP request.

OpenTelemetryAPI.extract — Function

extract(carrier, [global_propagator], [current_context])

Extracts the value from an incoming request. For example, from the headers of an HTTP request.

Trace

The relationship between trace provider, tracer, span context and span is depicted below:

┌────────────────────────────┐
│Span                        │
│                            │
│ tracer                     │
│   ┌──────────────────────┐ │
│   │Tracer                │ │
│   │                      │ │
│   │  provider            │ │
│   │   ┌────────────────┐ │ │
│   │   │    Abstract    │ │ │
│   │   │ TracerProvider │ │ │
│   │   └────────────────┘ │ │
│   │  instrumentation     │ │
│   │   ┌────────────────┐ │ │
│   │   │ name           │ │ │
│   │   │ version        │ │ │
│   │   └────────────────┘ │ │
│   │                      │ │
│   └──────────────────────┘ │
│ span_context               │
│   ┌──────────────────────┐ │
│   │ trace_id             │ │
│   │ span_id              │ │
│   │ is_remote            │ │
│   │ trace_flag           │ │
│   │ trace_state          │ │
│   └──────────────────────┘ │
│ parent_span_context        │
│ kind                       │
│ start_time                 │
│ end_time                   │
│ attributes                 │
│ links                      │
│ events                     │
│ status                     │
└────────────────────────────┘

In OpenTelemetryAPI.jl, only one AbstractTracerProvider (DummyTracerProvider) is provided and is set as the global_tracer_provider.

Span is immutable. Modifiable fields like status, end_time are set to Ref.

To add Events and Links, users can call push!(span, event_or_link).

The attributes in the Span is a DynamicAttrs, and can be updated with the syntax like span[key]=value.

end_time will be set by default after the call to with_span.

Always use set_status! to update the status of span.

Note

In some other languages, only a dummy span is defined in API and the concrete span is defined in SDK. Personally I prefer to defined it in API, otherwise we need to define many getter methods.

OpenTelemetryAPI.global_tracer_provider — Function

get the global tracer provider

set the global tracer provider to p

OpenTelemetryAPI.Tracer — Type

Tracer(;provider=global_tracer_provider(), instrumentation=InstrumentationInfo())

For instrumentation library developers, instrumentation must be configured clearly instead of the default value.

OpenTelemetryAPI.Span — Type

Span(name::String, tracer::Tracer{TracerProvider};kw...)

Keyword Arguments

context=current_context(),

kind=SPAN_KIND_INTERNAL,

attributes=Dict{String, TAttrVal}(),

links=Link[],

events=Event[],

start_time::UInt=time()*10^9, the nanoseconds.

is_remote=false

OpenTelemetryAPI.current_span — Function

current_span([current_context])

Get the span in the current context.

OpenTelemetryAPI.with_span — Function

with_span(f, s::AbstractSpan;kw...)

Call function f with the current span set to s.

Keyword arguments

end_on_exit=true, controls whether to call end! after f or not.

record_exception=true, controls whether to record the exception.

set_status_on_exception=true, decides whether to set status to SPAN_STATUS_ERROR automatically when an exception is caught.

Base.push! — Method

A specialized variant of add_event! to record exceptions. Usually used in a try... catch...end to capture the backtrace. If the ex is rethrowed in the catch...end, is_rethrow_followed should be set to true.

Missing docstring.

Missing docstring for Base.push!(s::Span, link::Link). Check Documenter's build log for details.

Base.setindex! — Method

(s::Span)[key] = val

Set the attributes in span s. Only valid when the span is not ended yet.

Missing docstring.

Missing docstring for Base.getindex(s::Span, key). Check Documenter's build log for details.

OpenTelemetryAPI.set_status! — Function

set_status!(s::Span, code::SpanStatusCode, description=nothing)

Update the status of span s by following the original specification. description is only considered when the code is SPAN_STATUS_ERROR. Only valid when the span is not ended yet.

OpenTelemetryAPI.end! — Function

end!(s::Span, t=UInt(time()*10^9))

Set the end time of the span and trigger span processors.

OpenTelemetryAPI.is_recording — Function

is_recording(s::Span)

Returns true if this span s is recording information like Event operations, attribute modification using setindex!, etc.

Some other exported symbols:

Missing docstring.

Missing docstring for TraceIdType. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SpanIdType. Check Documenter's build log for details.

Missing docstring.

Missing docstring for INVALID_TRACE_ID. Check Documenter's build log for details.

Missing docstring.

Missing docstring for INVALID_SPAN_ID. Check Documenter's build log for details.

Missing docstring.

Missing docstring for current_span. Check Documenter's build log for details.

Missing docstring.

Missing docstring for TraceFlag. Check Documenter's build log for details.

OpenTelemetryAPI.TraceState — Type

TraceState(entries::Pair{String,String}...)

TraceState carries vendor-specific trace identification data, represented as a list of key-value pairs. TraceState allows multiple tracing systems to participate in the same trace. It is fully described in the W3C Trace Context specification.

OpenTelemetryAPI.SpanContext — Type

SpanContext(;span_id, trace_id, is_remote, trace_flag=TraceFlag(), trace_state=TraceState())

A SpanContext represents the portion of a Span which must be serialized and propagated along side of a distributed context. SpanContexts are immutable.

The OpenTelemetry SpanContext representation conforms to the W3C TraceContext specification. It contains two identifiers - a TraceId and a SpanId - along with a set of common TraceFlags and system-specific TraceState values.

TraceId A valid trace identifier is a 16-byte array with at least one non-zero byte.

SpanId A valid span identifier is an 8-byte array with at least one non-zero byte.

TraceFlags contain details about the trace. Unlike TraceState values, TraceFlags are present in all traces. The current version of the specification only supports a single flag called sampled.

TraceState carries vendor-specific trace identification data, represented as a list of key-value pairs. TraceState allows multiple tracing systems to participate in the same trace. It is fully described in the W3C Trace Context specification.

Missing docstring.

Missing docstring for INVALID_SPAN_CONTEXT. Check Documenter's build log for details.

OpenTelemetryAPI.span_context — Function

span_context([s::Span])

Get the SpanContext from a span s. If s is not specified, current_span() will be used. nothing is returned if no span context found.

Missing docstring.

Missing docstring for SPAN_KIND_UNSPECIFIED. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SPAN_KIND_INTERNAL. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SPAN_KIND_SERVER. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SPAN_KIND_CLIENT. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SPAN_KIND_PRODUCER. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SPAN_KIND_CONSUMER. Check Documenter's build log for details.

OpenTelemetryAPI.LimitInfo — Type

LimitInfo(;kw...)

Used in TracerProvider to configure generated Tracer.

Keyword arguments:

span_attribute_count_limit::Int = 128

span_attribute_value_length_limit::Union{Nothing,Int} = nothing

span_event_count_limit::Int = 128

span_link_count_limit::Int = 128

OpenTelemetryAPI.Link — Type

Link(span_context, attributes)

See more details at links between spans.

OpenTelemetryAPI.Event — Type

Event(name, timestamp=time()*10^9, attributes=StaticAttrs())

timestamp is the nanoseconds.

Missing docstring.

Missing docstring for SPAN_STATUS_UNSET. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SPAN_STATUS_OK. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SPAN_STATUS_ERROR. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SpanStatu. Check Documenter's build log for details.

Metric

The relationship between MeterProvider, Meter and different instruments are depicted below:

 ┌─────────────────────────────┐
 │AbstractInstrument           │
 │                             │
 │  name                       │
 │  unit                       │
 │  description                │
 │                             │
 │  meter                      │
 │   ┌───────────────────────┐ │
 │   │Meter                  │ │
 │   │                       │ │
 │   │  provider             │ │
 │   │   ┌────────────────┐  │ │
 │   │   │   Abstract     │  │ │
 │   │   │ MeterProvider  │  │ │
 │   │   └────────────────┘  │ │
 │   │  name                 │ │
 │   │  version              │ │
 │   │  schema_url           │ │
 │   │                       │ │
 │   │  instrumentation      │ │
 │   │   ┌────────────────┐  │ │
 │   │   │ name           │  │ │
 │   │   │ version        │  │ │
 │   │   └────────────────┘  │ │
 │   │  instruments          │ │
 │   │                       │ │
 │   │    * Counter          │ │
 │   │    * Histogram        │ │
 │   │    * UpDownCounter    │ │
 │   │    * ObservableCounter│ │
 │   │    * Observable       │ │
 │   │      UpDownCounter    │ │
 │   └───────────────────────┘ │
 │                             │
 └─────────────────────────────┘

An Instrument belongs to a Meter, each Meter may contain many different Instruments. Similarly, a Meter belongs to a MeterProvider and a MeterProvider may contain many different Meters.

Missing docstring.

Missing docstring for global_meter_provider. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Meter. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Measurement. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Counter. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ObservableCounter. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Histogram. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ObservableGauge. Check Documenter's build log for details.

Missing docstring.

Missing docstring for UpDownCounter. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ObservableUpDownCounter. Check Documenter's build log for details.

SDK

Two common exporters are provided to for debugging:

Missing docstring.

Missing docstring for InMemoryExporter. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ConsoleExporter. Check Documenter's build log for details.

Trace

In SDK, a dedicated TraceProvider is provided.

Missing docstring.

Missing docstring for TraceProvider. Check Documenter's build log for details.

OpenTelemetrySDK.CompositSpanProcessor — Type

CompositSpanProcessor(processors...)

A wrapper of different concrete span processors. Users can also push! new span processors into it after construction.

See also SimpleSpanProcessor.

Missing docstring.

Missing docstring for ALWAYS_ON. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ALWAYS_OFF. Check Documenter's build log for details.

OpenTelemetrySDK.DEFAULT_ON — Constant

Sampler that respects its parent span's sampling decision, but otherwise always samples.

OpenTelemetrySDK.DEFAULT_OFF — Constant

Sampler that respects its parent span's sampling decision, but otherwise never samples.

Missing docstring.

Missing docstring for TraceIdRatioBased. Check Documenter's build log for details.

Metric

The current implementation of metrics in SDK is mainly inspired by the dotnet sdk.

┌──────────────────────────────────────────┐
│MeterProvider                             │
│                                          │
│  meters                                  │
│  views                                   │
│                                          │
│  instrument_associated_metric_names      │
│    instrument =>  Set{metric_name}       │
│                                          │
│  metrics                                 │
│    name => metric                        │
│    ┌───────────────────────────────────┐ │
│    │Metric                             │ │
│    │                                   │ │
│    │  name                             │ │
│    │  description                      │ │
│    │  criteria                         │ │
│    │  aggregation                      │ │
│    │    ┌──────────────────────────┐   │ │
│    │    │AggregationStore          │   │ │
│    │    │                          │   │ │
│    │    │  attributes => data_point│   │ │
│    │    │   ┌─────────────────┐    │   │ │
│    │    │   │AbstractDataPoint│    │   │ │
│    │    │   │                 │    │   │ │
│    │    │   │  value          │    │   │ │
│    │    │   │  start_time     │    │   │ │
│    │    │   │  end_time       │    │   │ │
│    │    │   │  exemplars      │    │   │ │
│    │    │   │ ┌────────────┐  │    │   │ │
│    │    │   │ │Exemplar    │  │    │   │ │
│    │    │   │ │            │  │    │   │ │
│    │    │   │ │ value      │  │    │   │ │
│    │    │   │ │ trace_id   │  │    │   │ │
│    │    │   │ │ span_id    │  │    │   │ │
│    │    │   │ └────────────┘  │    │   │ │
│    │    │   │                 │    │   │ │
│    │    │   └─────────────────┘    │   │ │
│    │    │                          │   │ │
│    │    └──────────────────────────┘   │ │
│    │                                   │ │
│    └───────────────────────────────────┘ │
│                                          │
└──────────────────────────────────────────┘

A View specifies which instruments are grouped together through Criteria. For each view, a Metric is created to store the Measurements. Each metric may have many different dimensions configured by StaticAttrs in a Measurement. For each dimension, we may also collect those [Exemplar]s in the mean while.

Design decisions

For each registered instrument, we have stored the associated metrics configured by views into the instrument_associated_metric_names field. So that for each pair of instrument => measurement, we can quickly determine which metrics to update.

The make sure that measurements with the same attribute key-values but with different order can be updated in the same dimension in the AggregationStore, a design from opentelemetry-dotnet#2374 is borrowed here.

Missing docstring.

Missing docstring for MeterProvider. Check Documenter's build log for details.

Missing docstring.

Missing docstring for View. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Metric. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AggregationStore. Check Documenter's build log for details.