OpenTelemetryAPI

The content in this page is organized in the same order as the OpenTelemetry Specification.

Context

All the MUST items in the original specification are implemented.

Context is implemented as a wrapper of NamedTuple, which means it is immutable. Each Task has exactly ONE Context instance, which is injected into the task_local_storage of the current_task by the parent task automatically.

Warning

Type piracy is used to the propagate context between tasks. See more discussions in #32

create_key is used to create a context key. But it is not exported yet because it seems to be only used internally until now.

Base.getindex(::Context, key) is implemented so to get a value in a Context, one can simply call ctx[key] to get the associated value of a key in a ctx.

Setting value of a Context is not directly supported. Given that Context is immutable, updating an immutable object in Julia seems strange. We provide the with_context function to create a new context based on the key-value pairs in the current_context. This syntax is more common than the attach/detach operations in the original specification.

OpenTelemetryAPI.current_context — Method

Return the Context associated with the caller's current execution unit.

OpenTelemetryAPI.with_context — Method

with_context(f, [context]; kv...)

Run function f in the context. If extra kv pairs are provided, they will be merged with the context to form a new context. When context is not provided, the current_context will be used.

Propagators

inject! and extract are provided based on the original specification. However, TextMapPropagator is not implemented yet! Personally I feel that every propagator may depends on a third party package. To minimize the dependencies of OpenTelemetryAPI.jl, those specialized propagators can be registered as independent packages.

The GLOBAL_PROPAGATOR is set to a CompositePropagator, with multiple dispatch, each inner propagator can be customized to handle different contexts and carriers. Since it's mainly used internally for now, it's not exposed yet.

OpenTelemetryAPI.extract — Function

extract(carrier, [global_propagator], [current_context])

Extracts the value from an incoming request. For example, from the headers of an HTTP request.

OpenTelemetryAPI.inject! — Function

inject(carrier, [global_propagator], [current_context])

Injects the value into a carrier. For example, into the headers of an HTTP request.

Trace

The relationship between trace provider, tracer, span context and span is depicted below:

┌────────────────────────────┐
│ AbstractSpan               │
│   ┌──────────────────────┐ │
│   │ Tracer               │ │
│   │   ┌────────────────┐ │ │
│   │   │    Abstract    │ │ │
│   │   │ TracerProvider │ │ │
│   │   └────────────────┘ │ │
│   │  instrumentation     │ │
│   │   ┌────────────────┐ │ │
│   │   │ name           │ │ │
│   │   │ version        │ │ │
│   │   └────────────────┘ │ │
│   └──────────────────────┘ │
└────────────────────────────┘

In OpenTelemetryAPI.jl, only one concrete AbstractTracerProvider (the DummyTracerProvider) is provided. It is set as the default global_tracer_provider. Without SDK installed, the with_span will only create a NonRecordingSpan. This is to make the OpenTelemetryAPI.jl lightweight enough so that instrumentation package users can happily add it as a dependency without losing performance.

OpenTelemetryAPI.AbstractSpan — Type

Each concrete span should have the following interfaces implemented.

tracer

span_context

span_kind

parent_span_context

attributes

is_recording

start_time

end_time

end_span!

span_status!

span_status

span_name!

span_name

span_links

span_events

OpenTelemetryAPI.AbstractTracerProvider — Type

A tracer provider is a part of an Tracer. For each concrete tracer provider, resource and OpenTelemetryAPI.create_span(name::String, tracer::Tracer{<:YourCustomProvider}) should also be implemented.

OpenTelemetryAPI.Tracer — Type

Tracer(name="Main", version=v"0.0.1-dev";provider=global_tracer_provider())

The name and version will form the InstrumentationInfo. value.

OpenTelemetryAPI.end_span! — Method

end_span!([s=current_span()], [t=UInt(time()*10^9)])

Set the end time of the span and trigger span processors. Note t is the nanoseconds.

OpenTelemetryAPI.global_tracer_provider! — Method

global_tracer_provider!(p)

Set the global tracer provider to p.

OpenTelemetryAPI.global_tracer_provider — Method

global_tracer_provider()

Get the global tracer provider.

OpenTelemetryAPI.is_recording — Method

is_recording([current_span()])

Returns true if this span s is recording information like Event operations, attribute modification using setindex!, etc.

OpenTelemetryAPI.parent_span_context — Method

parent_span_context(s::AbstractSpan)

Get the SpanContext from parent span.

OpenTelemetryAPI.provider — Method

provider(s::AbstractSpan)

Get the AbstractTracerProvider which generates the tracer that the span s resides in.

OpenTelemetryAPI.resource — Method

resource(s::AbstractSpan)

Get the associated resource of the span s. Fall back to resource(provider(::AbstractSpan))

OpenTelemetryAPI.span_context — Method

span_context([s::AbstractSpan])

Get the SpanContext from a span s. If s is not specified, current_span() will be used. nothing is returned if no span context found.

OpenTelemetryAPI.span_events — Method

span_events(s::AbstractSpan)

Get the recorded events in a span. A Limited is expected.

OpenTelemetryAPI.span_kind — Method

span_kind([current_span()])

Return SpanKind

OpenTelemetryAPI.span_links — Method

span_links(s::AbstractSpan)

Get the recorded links in a span. A Limited is expected.

OpenTelemetryAPI.span_name! — Method

span_name!([current_span()], name::String)

OpenTelemetryAPI.span_name — Method

span_name([current_span()])

OpenTelemetryAPI.span_status! — Method

span_status!([current_span()], code::SpanStatusCode, description=nothing)

Update the status of span s by following the original specification. description is only considered when the code is SPAN_STATUS_ERROR. Only valid when the span is not ended yet.

OpenTelemetryAPI.span_status — Method

span_status([current_span()])

Get status of the span. A SpanStatusCode is returned.

OpenTelemetryAPI.tracer — Method

tracer(s::AbstractSpan)

Get the Tracer which generates the span s.

OpenTelemetryAPI.with_span — Function

with_span(f, name::String, [tracer=Tracer()];kw...)

Call function f with the current span set a newly created one of name with tracer.

Keyword arguments

end_on_exit=true, controls whether to call end_span! after f or not.

record_exception=true, controls whether to record the exception.

set_status_on_exception=true, decides whether to set status to SPAN_STATUS_ERROR automatically when an exception is caught.

The rest keyword arguments are forwarded to create_span.

Metric

The relationship between MeterProvider, Meter and different instruments are depicted below:

 ┌─────────────────────────────┐
 │AbstractInstrument           │
 │                             │
 │  name                       │
 │  unit                       │
 │  description                │
 │                             │
 │  meter                      │
 │   ┌───────────────────────┐ │
 │   │Meter                  │ │
 │   │                       │ │
 │   │  provider             │ │
 │   │   ┌────────────────┐  │ │
 │   │   │   Abstract     │  │ │
 │   │   │ MeterProvider  │  │ │
 │   │   └────────────────┘  │ │
 │   │  name                 │ │
 │   │  version              │ │
 │   │  schema_url           │ │
 │   │                       │ │
 │   │  instrumentation      │ │
 │   │   ┌────────────────┐  │ │
 │   │   │ name           │  │ │
 │   │   │ version        │  │ │
 │   │   └────────────────┘  │ │
 │   │  instruments          │ │
 │   │                       │ │
 │   │    * Counter          │ │
 │   │    * Histogram        │ │
 │   │    * UpDownCounter    │ │
 │   │    * ObservableCounter│ │
 │   │    * Observable       │ │
 │   │      UpDownCounter    │ │
 │   └───────────────────────┘ │
 │                             │
 └─────────────────────────────┘

An Instrument belongs to a Meter, each Meter may contain many different Instruments. Similarly, a Meter belongs to a MeterProvider and a MeterProvider may contain many different Meters.

OpenTelemetryAPI.AbstractInstrument — Type

AbstractInstrumentis the super type of all instruments, which are used to report [Measurement`](@ref)s.

See also the specification:

OpenTelemetryAPI.AbstractMeterProvider — Type

A meter provider defines how to collect and update Measurements. Each meter provider should have the following interfaces implemented:

Base.push!(provider, m::Meter), register a meter.

Base.push!(provider, ins::AbstractInstrument), register an instrument.

Base.push!(provider, (ins::AbstractInstrument, m::Measurement)), update a measurement.

OpenTelemetryAPI.Meter — Type

Meter(name::String;kw...)

Meter is responsible for creating instruments.

Keyword Arguments:

provider::P = global_meter_provider()

version = v"0.0.1-dev"

schema_url = ""

OpenTelemetryAPI.global_meter_provider! — Method

global_meter_provider!(p)

Set the global meter provider to p.

OpenTelemetryAPI.global_meter_provider — Method

global_meter_provider()

Get the global meter provider.

Instruments

All instruments provided here can be classified into two categories: AbstractSyncInstrument and AbstractAsyncInstrument.

OpenTelemetryAPI.AbstractAsyncInstrument — Type

Async instrument usually has a callback function (which is named Observable in OpenTelemetry) to get its measurement.

Note

If the return of the callback function is not a Measurement, it will be converted into a Measurement with an empty StaticAttrs implicitly when being uploaded to the associated meter provider.

OpenTelemetryAPI.AbstractSyncInstrument — Type

Sync instrument usually pass its measurement immediately.

OpenTelemetryAPI.Counter — Type

Counter{T}(name, meter;unit="", description="") where T

Counter is a AbstractSyncInstrument which supports non-negative increments.

See more details from the specification.

OpenTelemetryAPI.Histogram — Type

Histogram{T}(name, meter; unit = "", description = "") where {T}

HistogramisAbstractSyncInstrument` which can be used to report arbitrary values that are likely to be statistically meaningful. It is intended for statistics such as histograms, summaries, and percentile.

See more details from the specification

OpenTelemetryAPI.Measurement — Type

Measurement(value, [attributes=StaticAttrs()])

This is to follow the specification

See also StaticAttrs

OpenTelemetryAPI.ObservableCounter — Type

ObservableCounter{T}(callback, name, meter; unit = "", description = "") where {T}

ObservableCounter is an AbstractAsyncInstrument which reports monotonically increasing value(s) when the instrument is being observed.

See more details from the specification

OpenTelemetryAPI.ObservableGauge — Type

ObservableGauge{T}(callback, name, meter; unit = "", description = "",) where {T}

ObservableGauge is an AbstractAsyncInstrument which reports non-additive value(s) (e.g. the room temperature - it makes no sense to report the temperature value from multiple rooms and sum them up) when the instrument is being observed.

See also the details from the specification

OpenTelemetryAPI.ObservableUpDownCounter — Type

ObservableUpDownCounter{T}(callback, name, meter; unit = "", description = "") where {T}

ObservableUpDownCounter is an AbstractAsyncInstrument which reports additive value(s) (e.g. the process heap size - it makes sense to report the heap size from multiple processes and sum them up, so we get the total heap usage) when the instrument is being observed.

See more details from the specification.

OpenTelemetryAPI.UpDownCounter — Type

UpDownCounter{T}(name, meter; unit = "", description = "") where {T}

UpDownCounter is a AbstractSyncInstrument which supports increments and decrements.

See also the details from the specification.

Logging

The idea is simple, a OtelLogTransformer is provided to transform each logging message into a LogRecord. To understand how to use it, users should be familiar with how TransformerLogger from LoggingExtras.jl works.

OpenTelemetryAPI.LogRecord — Type

LogRecord(;kw...)

A Julia representation of the Log Data Model.

OpenTelemetryAPI.OtelLogTransformer — Type

OtelLogTransformer(resource::Resource)

It can be used as a function f to the TransformerLogger. After applying this transformer, a LogRecord will be returned.

Misc

OpenTelemetryAPI.TAttrVal — Type

Valid type of attribute value.

OpenTelemetryAPI.CompositePropagator — Type

CompositePropagator(propagators::Vector)

OpenTelemetryAPI.DynamicAttrs — Type

DynamicAttrs(attrs::Dict{String, TAttrVal};count_limit=128, value_length_limit=nothing)
DynamicAttrs(attrs::Pair{String, TAttrVal}...;count_limit=128, value_length_limit=nothing)

Here we use a Dict internally to represent the mutable version of the Attributes specification. If the value_length_limit is set to a positive int, the value of String or each element in a value of Vector{String} will be truncated to a maximum length of value_length_limit. By default we do not do the truncation. When adding new pairs into it, if the number of attributes exceeds the count_limit, it will be dropped. You can get the number of dropped pairs via n_dropped.

See also StaticAttrs.

OpenTelemetryAPI.Event — Type

Event(name, timestamp=time()*10^9, attributes=StaticAttrs())

timestamp is the nanoseconds.

OpenTelemetryAPI.InstrumentationInfo — Type

InstrumentationInfo(;name="Main", version=v"0.0.1-dev")

Usually used in an instrumentation package.

OpenTelemetryAPI.LimitInfo — Type

LimitInfo(;kw...)

Used in TracerProvider to configure generated Tracer.

Keyword arguments:

span_attribute_count_limit::Int = 128

span_attribute_value_length_limit::Union{Nothing,Int} = nothing

span_event_count_limit::Int = 128

span_link_count_limit::Int = 128

OpenTelemetryAPI.Limited — Type

Limited(container; limit=32)

Create a container wrapper with limited elements. It supports the following common containers:

AbstractDict. If the length of the dict contains more elements than limit. Then pop! will be called repeatedly until the length is equal to limit. Further new key-value pair insertions will be ignored. If the key already exists in the dict, then the corresponding value will always be updated. n_dropped will return the number of dropped insertions.

AbstractVector. Similar to AbstractDict.

The following methods from Base are defined on Limited which are then forwarded to the inner container. Feel free to create a PR if you find any method you need is missing:

Base.getindex

Base.setindex!

Base.iterate

Base.length

Base.haskey

Base.push!. Only defined on containers of AbstractVector.

OpenTelemetryAPI.Link — Type

Link(span_context, attributes)

See more details at links between spans.

OpenTelemetryAPI.Resource — Type

Resource(;attributes=StaticAttrs(), schema_url="")

Quoted from the specification:

Resource captures information about the entity for which telemetry is recorded. For example, metrics exposed by a Kubernetes container can be linked to a resource that specifies the cluster, namespace, pod, and container name.

Resource may capture an entire hierarchy of entity identification. It may describe the host in the cloud and specific container or an application running in the process.

OpenTelemetryAPI.SpanContext — Type

SpanContext(;span_id, trace_id, is_remote, trace_flag=TraceFlag(), trace_state=TraceState())

A SpanContext represents the portion of a Span which must be serialized and propagated along side of a distributed context. SpanContexts are immutable.

The OpenTelemetry SpanContext representation conforms to the W3C TraceContext specification. It contains two identifiers - a TraceId and a SpanId - along with a set of common TraceFlags and system-specific TraceState values.

TraceId A valid trace identifier is a 16-byte array with at least one non-zero byte.

SpanId A valid span identifier is an 8-byte array with at least one non-zero byte.

TraceFlags contain details about the trace. Unlike TraceState values, TraceFlags are present in all traces. The current version of the specification only supports a single flag called sampled.

TraceState carries vendor-specific trace identification data, represented as a list of key-value pairs. TraceState allows multiple tracing systems to participate in the same trace. It is fully described in the W3C Trace Context specification.

OpenTelemetryAPI.SpanStatus — Type

SpanStatus(code, description=nothing)

Possible codes are:

SPAN_STATUS_UNSET

SPAN_STATUS_ERROR

SPAN_STATUS_OK

description is required when code is SPAN_STATUS_ERROR.

OpenTelemetryAPI.StaticAttrs — Type

StaticAttrs(attrs::NamedTuple; value_length_limit=nothing)
StaticAttrs(attrs::Pair{String, TAttrVal}...; value_length_limit=nothing)
StaticAttrs(attrs::Pair{Symbol, TAttrVal}...; value_length_limit=nothing)

Here we use the NamedTuple internally to efficiently represent the immutable version of the Attributes in the specification. If the value_length_limit is set to a positive int, the value of String or each element in a value of Vector{String} will be truncated to a maximum length of it. By default we do not do the truncation.

See also DynamicAttrs.

OpenTelemetryAPI.TraceState — Type

TraceState(entries::Pair{String,String}...)

TraceState carries vendor-specific trace identification data, represented as a list of key-value pairs. TraceState allows multiple tracing systems to participate in the same trace. It is fully described in the W3C Trace Context specification.

Base.getindex — Method

Base.getindex(s::AbstractSpan, key)

Look up key in the attributes of the span s.

Base.haskey — Method

Base.haskey(s::AbstractSpan, key)

Check if the span s has the key in its attributes.

Base.push! — Method

Base.push!([s::AbstractSpan], event::Event)

Add an Event into the span s.

Base.push! — Method

Base.push!([current_span()], link::Link)

Add a Link into the span s.

Base.push! — Method

Base.push!(s::AbstractSpan, ex::Exception; is_rethrow_followed = false)

A specialized variant of Event to record exceptions. Usually used in a try... catch...end to capture the backtrace. If the ex is rethrowed in the catch...end, is_rethrow_followed should be set to true.

Base.setindex! — Method

(s::AbstractSpan)[key] = val

Set the attributes in span s. Only valid when the span is not ended yet.

OpenTelemetryAPI.attributes — Method

attributes(s::AbstractSpan)

Return either StaticAttrs or DynamicAttrs in the span s.

OpenTelemetryAPI.create_span — Method

Here we follow the Behavior of the API in the absence of an installed SDK.

OpenTelemetryAPI.current_span — Method

current_span([current_context])

Get the span in the current context.

OpenTelemetryAPI.end_time — Method

end_time(s::AbstractSpan)

Get the end time of span s in nanoseconds.

OpenTelemetryAPI.n_dropped — Method

n_dropped(x::Limited)

Return the total number of dropped elements since creation.

OpenTelemetryAPI.start_time — Method

start_time(s::AbstractSpan)

Get the start time of span s in nanoseconds.