OpenTelemetryExporterPrometheus

In this package, only one AbstractExporter is exported: the PrometheusExporter. With this exporter, one can use the Prometheus to view the Metrics.

OpenTelemetryExporterPrometheus.PrometheusExporter — Type

PrometheusExporter(; host = "127.0.0.1", port = 9966, kw...)

It will setup a http server configured by host and port on initialization. Here the extra keyword arguments kw will be forwarded to HTTP.listen.

Usage

r = MetricReader(PrometheusExporter())

Note that since PrometheusExporter is pull based exporter, which means there's no need to execute r() to update the metrics.